first test

+++ src/main/java/com/eCommerce/service/impl/BookServiceImpl.java
@@ -33,12 +33,14 @@ public class BookServiceImpl implements BookService{
 	}
 
 	@Override
-	public void save(Book book) {
-		bookRepository.save(book);
+	public Book save(Book book) {
+		return bookRepository.save(book);
 	}
 
 	@Override
 	public void delete(Long id) {
+		Long count = bookRepository.countById(id);
+		if(count == null || count == 0) throw new RuntimeException("Could not find any book with ID:" + id);
 		bookRepository.deleteById(id);
 	}
